backend/requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchain_huggingface
langchain_groq
faiss-cpu
beautifulsoup4
python-dotenv